refactor: use standard names for method channel and platform interface packages
