+++ packages/wifi_scan/lib/wifi_scan_method_channel.dart
